Update feature news management and feedback management

+++ admin/quanlydangkyphong/xuly.php
@@ -1,4 +1,4 @@
-<?php 
+<?php
 session_start();
 include_once('../../config/database.php');
 
@@ -20,7 +20,7 @@ if(isset($_GET['action'])){
   				$ngay=$date['year']."-".$date['mon']."-".($date['mday']+3)." ".$date['hours'].":".$date['minutes'].":".$date['seconds'];
   				$td="Thông Báo Đăng Ký Phòng Ký Túc Xá";
   				$nd="Bạn đã đăng ký thành công ! Phòng : ".$map.". Vui lòng lên ký túc xá để thanh toán tiền phòng và nhận phòng trước ngày ".$ngay. ". Nếu không lên nhận phòng hệ thống sẽ hủy phòng của bạn và thêm bạn vào danh sách Xấu. Xin Cảm ơn !!!";
-  				
+
   				$sql2="INSERT INTO `thongbao`(`MaSV`, `MaNV`, `TieuDe`, `NoiDung`, `LoaiTB`) VALUES ($masv,'$manv','$td','$nd',N'đăng ký')";
   				$rs2=mysqli_query($conn,$sql2);echo $masv.$manv.$map.$ngay;
   				if($rs2){
@@ -29,44 +29,46 @@ if(isset($_GET['action'])){
             if($rs){
   					 header('location:../index.php?action=quanlydangkyphong&view=quanlydangkyphong');
             }
-  				}		
+  				}
   			}
 			break;
 		case 'huy':
 			$madk=$_GET['madk'];
 			$admin=$_SESSION['nv_admin'];$manv=$admin['MaNV'];
-			$sql="update phong set SoNguoiHienTai=(SoNguoiHienTai -1) where MaPhong=(select MaPhong from chitietdangky where MaDK=$madk)";
-			$rs=mysqli_query($conn,$sql);
-			if($rs){
-
-				$sql="update chitietdangky set MaNV = '$manv' where MaDK=$madk";
-				$rs=mysqli_query($conn,$sql);echo 'ok';
+			if (isset($_POST['lyDoHuy'])) {
+				$lyDoHuy = $_POST['lyDoHuy'];
+				$sql="update phong set SoNguoiHienTai=(SoNguoiHienTai -1) where MaPhong=(select MaPhong from chitietdangky where MaDK=$madk)";
+				$rs=mysqli_query($conn,$sql);
 				if($rs){
-				  echo "ok";
-				  $sql1="select * from chitietdangky where MaDK=$madk";
-				  $rs1=mysqli_query($conn,$sql1);
-				  $row1=mysqli_fetch_array($rs1);
-				  $masv=$row1['MaSV']; $manv=$row1['MaNV'];$map=$row1['MaPhong'];
-				  date_default_timezone_set('Asia/Ho_Chi_Minh');
-				  $date=getdate();
-				  $ngay=$date['year']."-".$date['mon']."-".($date['mday']+3)." ".$date['hours'].":".$date['minutes'].":".$date['seconds'];
-				  $td="Thông Báo Đăng Ký Phòng Ký Túc Xá";
-				  $nd="Đăng Ký Phòng của bạn đã bị hủy. vui lòng lên gặp Nhân viên Ký túc xá để biết thêm chi tiết. Xin cảm ơn !!!";
-				  
-				  $sql2="INSERT INTO `thongbao`(`MaSV`, `MaNV`, `TieuDe`, `NoiDung`, `LoaiTB`) VALUES ($masv,'$manv','$td','$nd',N'chuyển phòng')";
-				  $rs2=mysqli_query($conn,$sql2);echo $masv.$manv.$map.$ngay;
-				  if($rs2){
-					$sql="delete from chitietdangky where MaDK='$madk'";
-					$rs=mysqli_query($conn,$sql);
+					$sql="update chitietdangky set MaNV = '$manv' where MaDK=$madk";
+					$rs=mysqli_query($conn,$sql);echo 'ok';
 					if($rs){
-						header('location:../index.php?action=quanlydangkyphong&view=quanlydangkyphong');
+					  echo "ok";
+					  $sql1="select * from chitietdangky where MaDK=$madk";
+					  $rs1=mysqli_query($conn,$sql1);
+					  $row1=mysqli_fetch_array($rs1);
+					  $masv=$row1['MaSV']; $manv=$row1['MaNV'];$map=$row1['MaPhong'];
+					  date_default_timezone_set('Asia/Ho_Chi_Minh');
+					  $date=getdate();
+					  $ngay=$date['year']."-".$date['mon']."-".($date['mday']+3)." ".$date['hours'].":".$date['minutes'].":".$date['seconds'];
+					  $td="Thông Báo Đăng Ký Phòng Ký Túc Xá";
+					  $nd="Đăng Ký Phòng của bạn đã bị hủy. Lý do: ".$lyDoHuy.". Vui lòng lên gặp Nhân viên Ký túc xá để biết thêm chi tiết. Xin cảm ơn !!!";
+
+					  $sql2="INSERT INTO `thongbao`(`MaSV`, `MaNV`, `TieuDe`, `NoiDung`, `LoaiTB`) VALUES ($masv,'$manv','$td','$nd',N'chuyển phòng')";
+					  $rs2=mysqli_query($conn,$sql2);echo $masv.$manv.$map.$ngay;
+					  if($rs2){
+						$sql="delete from chitietdangky where MaDK='$madk'";
+						$rs=mysqli_query($conn,$sql);
+						if($rs){
+							header('location:../index.php?action=quanlydangkyphong&view=quanlydangkyphong');
+						}
+
+					  }
 					}
-	
-				  }   
 				}
 			}
 			break;
-		
+
 		default:
 			# code...
 			break;
